+++ linear.go
@@ -74,8 +74,8 @@ func newLinearGradient(colors []Color, positions []float64, mode BlendMode) Grad
 	}
 
 	return Gradient{
-		grad: gradbase,
-		dmin: positions[0],
-		dmax: positions[len(positions)-1],
+		Core: gradbase,
+		Min:  positions[0],
+		Max:  positions[len(positions)-1],
 	}
 }