+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - New `GradientBuilder.Reset()`
 - New `BlendLab`
+- `Gradient`'s fields is now public
+- `GradientCore` is now public
 
 ## v0.10.2